secure privkey
