+++ src-tauri/tauri.conf.json
@@ -41,6 +41,9 @@
 					"$VIDEO/*"
 				]
 			},
+			"path": {
+				"all": true
+			},
 			"shell": {
 				"all": false,
 				"open": true
