+++ src/app/auth/create/step-2.tsx
@@ -1,138 +1,126 @@
 import { useState } from 'react';
-import { useForm } from 'react-hook-form';
+import { Resolver, useForm } from 'react-hook-form';
 import { useNavigate } from 'react-router-dom';
 
-import { AvatarUploader } from '@shared/avatarUploader';
-import { BannerUploader } from '@shared/bannerUploader';
-import { LoaderIcon } from '@shared/icons';
-import { Image } from '@shared/image';
+import { EyeOffIcon, EyeOnIcon, LoaderIcon } from '@shared/icons';
 
-import { DEFAULT_AVATAR } from '@stores/constants';
 import { useOnboarding } from '@stores/onboarding';
+import { useStronghold } from '@stores/stronghold';
+
+import { useSecureStorage } from '@utils/hooks/useSecureStorage';
+
+type FormValues = {
+  password: string;
+};
+
+const resolver: Resolver<FormValues> = async (values) => {
+  return {
+    values: values.password ? values : {},
+    errors: !values.password
+      ? {
+          password: {
+            type: 'required',
+            message: 'This is required.',
+          },
+        }
+      : {},
+  };
+};
 
 export function CreateStep2Screen() {
   const navigate = useNavigate();
-  const createProfile = useOnboarding((state: any) => state.createProfile);
+  const setPassword = useStronghold((state) => state.setPassword);
 
-  const [picture, setPicture] = useState(DEFAULT_AVATAR);
-  const [banner, setBanner] = useState('');
+  const [pubkey, privkey] = useOnboarding((state) => [state.pubkey, state.privkey]);
+  const [passwordInput, setPasswordInput] = useState('password');
   const [loading, setLoading] = useState(false);
+
+  const { save } = useSecureStorage();
+
+  // toggle private key
+  const showPassword = () => {
+    if (passwordInput === 'password') {
+      setPasswordInput('text');
+    } else {
+      setPasswordInput('password');
+    }
+  };
 
   const {
     register,
+    setError,
     handleSubmit,
-    formState: { isDirty, isValid },
-  } = useForm();
+    formState: { errors, isDirty, isValid },
+  } = useForm<FormValues>({ resolver });
 
-  const onSubmit = (data: any) => {
+  const onSubmit = async (data: { [x: string]: string }) => {
     setLoading(true);
-    try {
-      const profile = {
-        ...data,
-        username: data.name,
-        display_name: data.name,
-        bio: data.about,
-      };
-      createProfile(profile);
+    if (data.password.length > 3) {
+      // add password to local state
+      setPassword(data.password);
+
+      // save privkey to secure storage
+      await save(pubkey, privkey, data.password);
+
       // redirect to next step
-      setTimeout(() => navigate('/auth/create/step-3', { replace: true }), 1200);
-    } catch {
-      console.log('error');
+      navigate('/auth/create/step-3', { replace: true });
+    } else {
+      setLoading(false);
+      setError('password', {
+        type: 'custom',
+        message: 'Password is required and must be greater than 3',
+      });
     }
   };
 
   return (
     <div className="mx-auto w-full max-w-md">
       <div className="mb-8 text-center">
-        <h1 className="text-xl font-semibold text-zinc-100">Create your profile</h1>
+        <h1 className="text-xl font-semibold text-zinc-100">
+          Set password to secure your key
+        </h1>
       </div>
-      <div className="w-full overflow-hidden rounded-xl border-t border-zinc-800/50 bg-zinc-900">
-        <form onSubmit={handleSubmit(onSubmit)} className="mb-0 flex flex-col">
-          <input
-            type={'hidden'}
-            {...register('picture')}
-            value={picture}
-            className="shadow-input relative h-10 w-full rounded-lg border border-black/5 px-3 py-2 shadow-black/5 !outline-none placeholder:text-zinc-400 dark:bg-zinc-800 dark:text-zinc-100 dark:shadow-black/10 dark:placeholder:text-zinc-500"
-          />
-          <input
-            type={'hidden'}
-            {...register('banner')}
-            value={banner}
-            className="shadow-input relative h-10 w-full rounded-lg border border-black/5 px-3 py-2 shadow-black/5 !outline-none placeholder:text-zinc-400 dark:bg-zinc-800 dark:text-zinc-100 dark:shadow-black/10 dark:placeholder:text-zinc-500"
-          />
-          <div className="relative">
-            <div className="relative h-44 w-full bg-zinc-800">
-              <Image
-                src={banner}
-                fallback="https://void.cat/d/QY1myro5tkHVs2nY7dy74b.jpg"
-                alt="user's banner"
-                className="h-full w-full object-cover"
+      <div className="flex flex-col gap-4">
+        <form onSubmit={handleSubmit(onSubmit)} className="flex flex-col gap-3">
+          <div className="flex flex-col gap-1">
+            <div className="relative">
+              <input
+                {...register('password', { required: true })}
+                type={passwordInput}
+                className="relative w-full rounded-lg bg-zinc-800 py-3 pl-3.5 pr-11 text-zinc-100 !outline-none placeholder:text-zinc-400"
               />
-              <div className="absolute left-1/2 top-1/2 z-10 h-full w-full -translate-x-1/2 -translate-y-1/2 transform">
-                <BannerUploader setBanner={setBanner} />
-              </div>
+              <button
+                type="button"
+                onClick={() => showPassword()}
+                className="group absolute right-2 top-1/2 -translate-y-1/2 transform rounded p-1 hover:bg-zinc-700"
+              >
+                {passwordInput === 'password' ? (
+                  <EyeOffIcon
+                    width={20}
+                    height={20}
+                    className="text-zinc-500 group-hover:text-zinc-100"
+                  />
+                ) : (
+                  <EyeOnIcon
+                    width={20}
+                    height={20}
+                    className="text-zinc-500 group-hover:text-zinc-100"
+                  />
+                )}
+              </button>
             </div>
-            <div className="mb-5 px-4">
-              <div className="relative z-10 -mt-7 h-14 w-14">
-                <Image
-                  src={picture}
-                  fallback={DEFAULT_AVATAR}
-                  alt="user's avatar"
-                  className="h-14 w-14 rounded-lg object-cover ring-2 ring-zinc-900"
-                />
-                <div className="absolute left-1/2 top-1/2 z-10 h-full w-full -translate-x-1/2 -translate-y-1/2 transform">
-                  <AvatarUploader setPicture={setPicture} />
-                </div>
-              </div>
+            <div className="text-sm text-zinc-500">
+              <p>
+                Password is use to secure your key store in local machine, when you move
+                to other clients, you just need to copy your private key as nsec or
+                hexstring
+              </p>
             </div>
+            <span className="text-sm text-red-400">
+              {errors.password && <p>{errors.password.message}</p>}
+            </span>
           </div>
-          <div className="flex flex-col gap-4 px-4 pb-4">
-            <div className="flex flex-col gap-1">
-              <label
-                htmlFor="name"
-                className="text-sm font-semibold uppercase tracking-wider text-zinc-400"
-              >
-                Name *
-              </label>
-              <input
-                type={'text'}
-                {...register('name', {
-                  required: true,
-                  minLength: 4,
-                })}
-                spellCheck={false}
-                className="relative h-10 w-full rounded-lg bg-zinc-800 px-3 py-2 text-zinc-100 !outline-none placeholder:text-zinc-500"
-              />
-            </div>
-            <div className="flex flex-col gap-1">
-              <label
-                htmlFor="about"
-                className="text-sm font-semibold uppercase tracking-wider text-zinc-400"
-              >
-                Bio
-              </label>
-              <textarea
-                {...register('about')}
-                spellCheck={false}
-                className="relative h-20 w-full resize-none rounded-lg bg-zinc-800 px-3 py-2 text-zinc-100 !outline-none placeholder:text-zinc-500"
-              />
-            </div>
-            <div className="flex flex-col gap-1">
-              <label
-                htmlFor="website"
-                className="text-sm font-semibold uppercase tracking-wider text-zinc-400"
-              >
-                Website
-              </label>
-              <input
-                type={'text'}
-                {...register('website', {
-                  required: false,
-                })}
-                spellCheck={false}
-                className="relative h-10 w-full rounded-lg bg-zinc-800 px-3 py-2 text-zinc-100 !outline-none placeholder:text-zinc-500"
-              />
-            </div>
+          <div className="flex items-center justify-center">
             <button
               type="submit"
               disabled={!isDirty || !isValid}
