+++ src/libs/storage.tsx
@@ -38,14 +38,13 @@ export async function getAccounts() {
 export async function createAccount(
   npub: string,
   pubkey: string,
-  privkey: string,
   follows?: string[][],
   is_active?: number
 ) {
   const db = await connect();
   const res = await db.execute(
     'INSERT OR IGNORE INTO accounts (npub, pubkey, privkey, follows, is_active) VALUES (?, ?, ?, ?, ?);',
-    [npub, pubkey, privkey, follows || '', is_active || 0]
+    [npub, pubkey, 'privkey is stored in secure storage', follows || '', is_active || 0]
   );
   if (res) {
     await createBlock(
